push

+++ src/main/java/Purchase/PurchaseCancelController.java
@@ -9,7 +9,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import dao.PurchaseInfoDao;
+import vo.PurchaseInfo;
 
+// 구매 내역에서 상품을 삭제했을 경우 이동하는 controller
 @WebServlet("/purchase/cancel")
 public class PurchaseCancelController extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
@@ -18,14 +20,23 @@ public class PurchaseCancelController extends HttpServlet {
 			return;
 		}
 		
+		if(request.getParameter("userIdx") == null) {
+			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+			return;
+		}
+		
 		int purchaseIdx = Integer.parseInt(request.getParameter("purchaseIdx"));
 		
+		int userIdx = Integer.parseInt(request.getParameter("userIdx"));
+		
 		PurchaseInfoDao dao = new PurchaseInfoDao();
+		
 		boolean result = dao.deletePurchaseIdx(purchaseIdx);
 		
 		if(result) {
-			// 200 상태 코드 전달
-			response.setStatus(HttpServletResponse.SC_OK);
+			response.sendRedirect("http://localhost/temperShop/purchase/history?userIdx="+userIdx);
+		} else {
+			
 		}
 	}
